Modularize various classes and enums
