+++ __tests__/pages/client/index.test.tsx
@@ -1,18 +1,20 @@
 // client/index.test.js
 import '@testing-library/jest-dom';
-import { render, screen, waitFor, within } from '@testing-library/react';
-import Client from '../../../pages/client';
-import { ClientTicket, ServiceCategory, clientService } from '../../../src/services/ClientService';
+import {
+  render, screen, waitFor,
+} from '@testing-library/react';
+import { ServiceCategory, ClientTicket } from '@lib';
 import { useRouter } from 'next/navigation';
+import Client from '../../../pages/client';
+import { clientService } from '../../../src/services/ClientService';
 
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn().mockImplementation(() => ({
     push: jest.fn(),
-  }))
+  })),
 }));
 
 describe('Client', () => {
-
   it('renders without crashing', () => {
     render(<Client />);
     expect(screen.getByText('Clients')).toBeInTheDocument();
@@ -23,9 +25,9 @@ describe('Client', () => {
       writable: true,
       value: { search: '?ticketNo=t123' },
     });
-    const tixSpy = jest.spyOn(clientService, 'getTicket')
-    const catSpy = jest.spyOn(clientService, 'getServiceCategories')
-    const statSpy = jest.spyOn(clientService, 'getServiceStatuses')
+    const tixSpy = jest.spyOn(clientService, 'getTicket');
+    const catSpy = jest.spyOn(clientService, 'getServiceCategories');
+    const statSpy = jest.spyOn(clientService, 'getServiceStatuses');
 
     render(<Client />);
 
@@ -39,12 +41,12 @@ describe('Client', () => {
   it('click breadcrumb', () => {
     const push = jest.fn();
     (useRouter as jest.Mock).mockImplementation(() => ({
-      push
+      push,
     }));
 
     render(<Client />);
     screen.getByText('Clients').click();
-    expect(push).toHaveBeenCalledWith('/clients')
+    expect(push).toHaveBeenCalledWith('/clients');
   });
 
   it('render category drop down', async () => {
@@ -55,7 +57,7 @@ describe('Client', () => {
 
     const cats = [
       new ServiceCategory({ id: 'CAT101', name: 'PET ADOPTION' }),
-      new ServiceCategory({ id: 'CAT102', name: 'Pet Fostering' })]
+      new ServiceCategory({ id: 'CAT102', name: 'Pet Fostering' })];
     const tix = new ClientTicket({});
 
     const tixSpy = jest.spyOn(clientService, 'getTicket')
@@ -64,7 +66,6 @@ describe('Client', () => {
     const catSpy = jest.spyOn(clientService, 'getServiceCategories')
       .mockReturnValue(Promise.resolve(cats));
 
-
     render(<Client />);
     expect(tixSpy).toHaveBeenCalledTimes(1);
     expect(catSpy).toHaveBeenCalledTimes(1);
@@ -72,9 +73,9 @@ describe('Client', () => {
     await waitFor(() => {
       // I'd look for a real text here that is renderer when the data loads
       expect(screen.getByText('Service Category')).toBeDefined();
-    })
+    });
     const drop = screen.getByTitle('Service category options');
-    expect(drop).toBeInTheDocument()
+    expect(drop).toBeInTheDocument();
 
     // selecting a dropdown will have to be covered later (not obvious how to do it)
   });
