+++ pages/clients/index.tsx
@@ -9,8 +9,9 @@ import { Toast } from 'primereact/toast';
 import { Toolbar } from 'primereact/toolbar';
 import React, { useEffect, useRef, useState } from 'react';
 import { useRouter } from 'next/navigation';
+import { ClientTicket } from '@lib';
 import ClientDialog from '../../src/components/ClientDialog';
-import { ClientTicket, clientService } from '../../src/services/ClientService';
+import { clientService } from '../../src/services/ClientService';
 
 function Clients() {
   const { push } = useRouter();
