+++ src/lib/ChangeLog.ts
@@ -1,0 +1,13 @@
+export default class ChangeLog {
+  date: Date;
+
+  representative: string;
+
+  description: string;
+
+  constructor(data: any) {
+    this.date = data.date ? data.date : new Date();
+    this.description = data.description;
+    this.representative = data.representative;
+  }
+}
