+++ src/lib/ClientRequest.ts
@@ -1,0 +1,35 @@
+import { RequestType, TicketType } from '@lib';
+
+export default class NewClientRequest {
+  requestType: RequestType = RequestType.clientNew;
+
+  ticketNo: string;
+
+  type: TicketType;
+
+  name: string;
+
+  email: string;
+
+  phone: string;
+
+  summary: string;
+
+  description: string;
+
+  date: Date;
+
+  representative: string;
+
+  constructor(data: any) {
+    this.ticketNo = data.ticketNo;
+    this.type = data.type;
+    this.name = data.name;
+    this.email = data.email;
+    this.phone = data.phone;
+    this.summary = data.summary;
+    this.description = data.description;
+    this.representative = data.representative;
+    this.date = data.date ? data.date : new Date();
+  }
+}
