+++ src/lib/RequestType.ts
@@ -1,0 +1,8 @@
+enum RequestType {
+  clientNew = 'client-new',
+  clientUpdate = 'client-new',
+  animalNew = 'animal-new',
+  animalUpdate = 'animal-update',
+}
+
+export default RequestType;
