+++ src/lib/ServiceCategory.ts
@@ -1,0 +1,10 @@
+export default class ServiceCategory {
+  id: string;
+
+  name: string;
+
+  constructor(input: any) {
+    this.id = input.id;
+    this.name = input.name;
+  }
+}
