+++ src/lib/ServiceStatus.ts
@@ -1,0 +1,19 @@
+export default class ServiceStatus {
+  id: string;
+
+  code: string;
+
+  name: string;
+
+  constructor(input: any) {
+    this.id = input.id;
+    this.code = input.code;
+    this.name = input.name;
+  }
+}
+export const Statuses: ServiceStatus[] = [
+  new ServiceStatus({ name: 'New', code: 'new' }),
+  new ServiceStatus({ name: 'In-progress', code: 'update' }),
+  new ServiceStatus({ name: 'Close', code: 'closed' }),
+  new ServiceStatus({ name: 'Blocked', code: 'blocked' }),
+];
