+++ src/lib/TicketType.ts
@@ -1,0 +1,8 @@
+enum TicketType {
+  walkin = 'walk-in',
+  email = 'email',
+  phone = 'phone',
+  other = 'other',
+}
+
+export default TicketType;
