+++ src/lib/index.ts
@@ -1,0 +1,12 @@
+/**
+ * This is a "barrel file" https://basarat.gitbook.io/typescript/main-1/barrel
+ * Its purpose is to make imports more concise and easier to remember
+ * while allowing for modularization of the code into separate files
+ */
+export { default as RequestType } from './RequestType';
+export { default as TicketType } from './TicketType';
+export { default as ServiceCategory } from './ServiceCategory';
+export { default as ServiceStatus, Statuses } from './ServiceStatus';
+export { default as ClientRequest } from './ClientRequest';
+export { default as ClientTicket } from './ClientTicket';
+export { default as ChangeLog } from './ChangeLog';
